Neuro/youtube/requirements.txt
flask==2.2.3
youtube-transcript-api==0.6.0
google-generativeai==0.3.1
python-dotenv==1.0.0